+++ packages/experimental/scripts/generate/templates/DISPLAY_NAME.js
@@ -8,6 +8,7 @@
 import cx from 'classnames';
 import { node, string } from 'prop-types';
 import React from 'react';
+import { Canary } from '../_Canary';
 
 /**
  * TODO: @import(s) of carbon components
@@ -16,20 +17,29 @@ import React from 'react';
 /**
  * TODO: Add use of Carbon prefix if needed
  */
-import { pkgPrefix /*, carbonPrefix */ } from '../../global/js/settings';
+import { pkg /*, carbon */ } from '../../settings';
 
-const blockClass = `${pkgPrefix}--STYLE_NAME`;
+const blockClass = `${pkg.prefix}--STYLE_NAME`;
+const componentName = 'DISPLAY_NAME';
+
+// NOTE: SCSS is not imported directly here: it is rolled up separately.
 
 /**
  * TODO: Description.
  */
-export const DISPLAY_NAME = ({ children, className, ...rest }) => {
-  return (
-    <div className={cx(blockClass, className)} {...rest}>
-      {children}
-    </div>
-  );
-};
+export const DISPLAY_NAME = !pkg.isComponentEnabled(componentName)
+  ? // Return canary if not released or flag not set
+    () => <Canary component={componentName} />
+  : // Main component code...
+    ({ children, className, ...rest }) => {
+      return (
+        <div className={cx(blockClass, className)} {...rest}>
+          {children}
+        </div>
+      );
+    };
+
+DISPLAY_NAME.displayName = componentName; // displayName is used in preference to function.name by React
 
 // TODO: Prop type checking - https://www.npmjs.com/package/prop-types#usage
 DISPLAY_NAME.propTypes = {