+++ packages/experimental/src/components/ActionBar/ActionBar.js
@@ -11,12 +11,12 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import ReactResizeDetector from 'react-resize-detector';
 
-import { pkgPrefix /*, carbonPrefix */ } from '../../global/js/settings';
+import { pkg /*, carbon */ } from '../../settings';
 import { OverflowMenu, OverflowMenuItem } from 'carbon-components-react';
 import uuidv4 from '../../global/js/utils/uuidv4';
 import unwrapIfFragment from '../../global/js/utils/unwrap-if-fragment';
 
-const blockClass = `${pkgPrefix}-action-bar`;
+const blockClass = `${pkg.prefix}-action-bar`;
 
 export const ActionBar = ({
   children,