+++ packages/experimental/src/components/Card/Card.stories.js
@@ -12,9 +12,9 @@ import styles from './_storybook-styles.scss'; // import index in case more file
 import mdx from './Card.mdx';
 import { ArrowRight24, Cloud32 } from '@carbon/icons-react';
 import { AspectRatio } from 'carbon-components-react';
-import {
-  storybookPrefixCanary as storybookPrefix /* , storybookPrefixReleased */,
-} from '../../../config';
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'Card');
 
 export default {
   title: `${storybookPrefix}/Card`,