+++ packages/experimental/src/components/PageHeader/PageHeader-test.js
@@ -8,7 +8,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import { pkgPrefix /*, carbonPrefix */ } from '../../global/js/settings';
+import { pkg } from '../../settings';
 
 import {
   BreadcrumbItem,
@@ -94,16 +94,16 @@ describe('PageHeader', () => {
   test('renders an empty header when no props are set', () => {
     render(<PageHeader />);
 
-    const header = document.querySelector(`section.${pkgPrefix}-page-header`);
+    const header = document.querySelector(`section.${pkg.prefix}-page-header`);
 
     expect(header).not.toBeNull();
     expect(
-      header.classList.contains(`${pkgPrefix}-page-header--background`)
+      header.classList.contains(`${pkg.prefix}-page-header--background`)
     ).toBe(false);
     expect(header.classList.contains(classNames[0])).toBe(false);
     expect(header.classList.contains(classNames[1])).toBe(false);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--action-bar`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--action-bar`)
     ).toHaveLength(0);
     expect(
       document.querySelectorAll('span.page-header-test--available-space')
@@ -111,19 +111,19 @@ describe('PageHeader', () => {
     expect(screen.queryAllByTestId('breadcrumbitem')).toHaveLength(0);
     expect(screen.queryAllByTestId('tabs')).toHaveLength(0);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--page-actions`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--page-actions`)
     ).toHaveLength(0);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--subtitle`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--subtitle`)
     ).toHaveLength(0);
     expect(screen.queryByText(subtitle)).toBeNull();
     expect(screen.queryAllByTestId('tags')).toHaveLength(0);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--title`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--title`)
     ).toHaveLength(0);
     expect(screen.queryByText(title)).toBeNull();
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--title-icon`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--title-icon`)
     ).toHaveLength(0);
   });
 
@@ -144,44 +144,42 @@ describe('PageHeader', () => {
       />
     );
 
-    const header = document.querySelector(`section.${pkgPrefix}-page-header`);
-    console.log('Look here', header.innerHTML);
-
+    const header = document.querySelector(`section.${pkg.prefix}-page-header`);
     expect(header).not.toBeNull();
     expect(
-      header.classList.contains(`${pkgPrefix}-page-header--background`)
+      header.classList.contains(`${pkg.prefix}-page-header--background`)
     ).toBe(true);
     expect(header.classList.contains(classNames[0])).toBe(true);
     expect(header.classList.contains(classNames[1])).toBe(true);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--action-bar`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--action-bar`)
     ).toHaveLength(1);
     expect(
       document.querySelectorAll('span.page-header-test--available-space')
     ).toHaveLength(1);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--breadcrumb`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--breadcrumb`)
     ).toHaveLength(1);
     expect(screen.getAllByText(/Breadcrumb [1-3]/)).toHaveLength(3);
     expect(screen.queryAllByTestId('tabs')).toHaveLength(1);
     expect(screen.getAllByText(/Tab [1-4]/)).toHaveLength(4);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--page-actions`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--page-actions`)
     ).toHaveLength(2);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--subtitle`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--subtitle`)
     ).toHaveLength(1);
     expect(screen.getByText(subtitle).textContent).toEqual(subtitle);
     expect(screen.queryAllByTestId('tabs')).toHaveLength(1);
     expect(screen.getAllByText('A tag')).toHaveLength(4);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--title`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--title`)
     ).toHaveLength(1);
     expect(
-      document.querySelector(`.${pkgPrefix}-page-header--title`).textContent
+      document.querySelector(`.${pkg.prefix}-page-header--title`).textContent
     ).toEqual(title);
     expect(
-      document.querySelectorAll(`.${pkgPrefix}-page-header--title-icon`)
+      document.querySelectorAll(`.${pkg.prefix}-page-header--title-icon`)
     ).toHaveLength(1);
   });
 });